+++ src/validators/validator.ts
@@ -1,0 +1,6 @@
+import {FtlFile} from '../models/ftl-file';
+import {Diagnostic, TextDocument} from 'vscode';
+
+export interface Validator {
+    validateFile(file: FtlFile, document: TextDocument, diagnostics: Diagnostic[]): void;
+}