+++ CHANGELOG.md
@@ -72,3 +72,4 @@ All notable changes to the "ftlXml" extension will be documented in this file.
 ## 0.3.6
 - Remove 'min_level' as an attribute from 'choice'
 - fix bug with 'damage', system can be 'room' or 'random'
+- show info when an event is not used anywhere